+++ cli/sol/bin/cmd_local.ml
@@ -959,6 +959,7 @@ let cmd =
     [ infra_cmd
     ; Cmd_status.local_cmd
     ; Cmd_logs.local_cmd
+    ; Cmd_fn.local_cmd
     ; Cmd_rollback.local_cmd
     ; Cmd_migrate.local_cmd
     ; Cmd_releases.local_cmd